blk/kernel: Fix error code mapping in KernelDevice::read.
pread returns -1 upon error and stores the error code in errno, and thus
the wrong error was being passed into is_expected_ioerr. This is handled
correctly just a few lines down where we return -errno, so it was likely
just an oversight when adapting this logic from the aio codepath, where
the return code is indeed the errno.

This logic has been incorrect since it was introduced in 2018 via
8d802da7e5333b37d9b4e806524fe507b3c0040b.

Signed-off-by: Joshua Baergen <jbaergen@digitalocean.com>