mds: include last modification time and change attr of a snap realm node in snap trace reply
To handle older clients, introduce a cephfs feature bit. The MDS encodes a "new"
SNapRealmInfo structure for clients supporting this feature (which includes the
last modification timestamp and a version/change_attr for the snap realm). For
clients not advertising this feature, the MDS uses the existing SnapRealmInfo
structure for snapshot traces.

Signed-off-by: Venky Shankar <vshankar@redhat.com>