qa: enable kclient test for newop test
The kclient have already fix this. This will only enable the upstream
kclient with the testing branch, the downstream ones may not include
the fixing yet, so skip them for now.

The nautilus will only support the syntax v1. And for kclient there
is not need to do the upgrade.

Fixes: https://tracker.ceph.com/issues/57591
Signed-off-by: Xiubo Li <xiubli@redhat.com>