working -- see status note below
Note to future me: this branch ended up being a bit of a dead end.
Refactoring scrub as it is now seems to be harder than starting from
scratch.  The new plan is to create a new crimson implementation from
whole cloth with an eye to being able to port it back later.