doc/radosgw: remove pipes from s3select.rst
Remove strange formatting introduced in
f3bb7a3db07042657b646c6270ab5e08ca9ac6e9.

Note to the documentation team: This is not a line-edit. This commit
includes nothing but the removal of pipes added to the left of much of
the text in this file. Several future commits will line-edit this file
and correct its formatting.

Signed-off-by: Zac Dover <zac.dover@proton.me>