test/libcephfs: update delegation test assertions
Some of the libcephfs tests ensure that the delegations are recalled
as a result of certain open requests.

The issue is that we're expecting the "opened" flag to be false
after the delegation gets recalled, which is racy.

  /home/ubuntu/ceph/src/test/libcephfs/deleg.cc:392: Failure
  Expected equality of these values:
    opened.load()
      Which is: true
    false

We'll update those checks, asserting that "opened" is true after
the delegation breaker thread returns.

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>