common/TrackedOp: fix race updating description with proper lock
It is not safe to "cache" a member which may be changed by a racing thread.

This reworks the locking so we can do a light-weight check if the description
is already generated wihtout acquiring the heavier TrackedOp::lock. If it's not
available yet or it needs regenerated, then get the proper locks to generate
it.

Fixes: a6608030857fc992ff1499faaf63c2f43ccc6741
Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>