libcephfs/deleg: Wait for the delegation to be timedout
MDSs will allow multiple clients to do r/w at the same time and the
filelock will be in LOCK_MIX state. So the open() in a second mounter
in the open_breaker_func() thread won't guarantee that after it
exiting the delegation in the mounter in main thread has already
timedout.

Fixes: https://tracker.ceph.com/issues/62245
Signed-off-by: Xiubo Li <xiubli@redhat.com>