doc: do not recommend pulling cephadm from git
First off, these instructions are wrong because the file has been
renamed to cephadm.py, and will not succeed as written. Amusingly, due
to curl's default behavior (and my lack of access to a time machine) the
initial command will seem to have succeeded until the user discovers
that HTML delivered as part of the body of a 404 response is not an
executable script.

Furthermore, in preparation for Reef we updated the documentation in the
cephadm subsection to use the results of a build, rather than source
files, because reef now contains a "compiled" cephadm. Currently, the
contents of the compilation is essentially the same-old cephadm script
but that will no longer be the case in the very near future (see
e1940dad8202cf47b20257277847e8f1141a3fdb). We missed this general
section of the Ceph docs that repeats the instructions to copy a raw
python file out of the Ceph tree on GitHub.  Update the section to
contain a simplified version of the instruction found in
doc/cephadm/install.rst.

Signed-off-by: John Mulligan <jmulligan@redhat.com>