cephadm: remove duplicates when listing files in scan
Fix running tox env flake8 during a rebase/merge where git ls-files
may return duplicate file names.

An earlier version of this change used the `--deuplicate` option to `git
ls-files` in order to remove duplicate files listed when running the
tool during rebase/merge. Mysteriously, this seemed to break the `ceph
API tests` in the CI. A completely unrelated command in a tox file that,
AFAIK, is not executed in that test suite. They ways of git are
subtle and often sucky.

Signed-off-by: John Mulligan <jmulligan@redhat.com>