doc/glossary: link to "ceph clients" from entry
Link to the "Ceph Clients" section of doc/architecture.rst from the
"Ceph Clients" entry in the glossary. A glossary entry should be a short
summary of the topic with which it deals, and it should direct the
reader to further and more detailed reading if the reader is interested.
This does that.

Signed-off-by: Zac Dover <zac.dover@proton.me>