dokan: simple case insensitive emulation
While CephFS is case sensitive, Windows software commonly assume
that the filesystem is case insensitive.

In order to unblock certain use cases, a short term solution is to
simply normalize paths when creating or accessing files or
directories.

This change adds two ceph-dokan parameters:

* --case-insensitive: if set, paths are normalized
* --force-lowercase: normalized paths are converted to lowercase
                     instead of uppercase

This trivial implementation has some limitations:

* the original case is not preserved
  * we could later on use xattr to store the original name
* can't access existing files that have a different case
* handled at ceph-dokan level

The advantage is that it's simple, shouldn't impact performance
and doesn't require libcephfs or mds changes.

In the future, we may conider a more robust implementation
at the mds and/or libcephfs level.

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>