messages/MOSDRepScrubMap.h: removed unused scrub_map_bl
This member has actually never been used.  It was introduced in
42be51d2 along with MOSDRepScrubMap, but wasn't included in
encode or decode nor was it referenced anywhere else in the commit.
Presumably, it was introduced during developement to contain the
encoded scrub map, but get_data() ended up being used instead and
scrub_map_bl didn't get removed.

Signed-off-by: Samuel Just <sjust@redhat.com>