mds/scrub: enqueue all child frags for a given fragset
Problem:
fragsets sent over to replicas for scrubbing are simplified i.e. they
are bit representation of the lease common ancestors of the frags that
need scrubbing on that replica. A search operation of a frag in the
frasget often fails to match exactly with the frags delegated to the
replica causing the scrub item to infinitely be held in the scrub stack.

Solution:
Test if the frag in the fragset sent over to the replica contains the
delegated frag as a child to accept it for scrubbing.

Fixes: https://tracker.ceph.com/issues/62658
Signed-off-by: Milind Changire <mchangir@redhat.com>