cephadm: update test to avoid using exception handling as an assertion
The use of an exception as an assertion mostly works but has the side
effect of hiding other errors. Hiding these errors can make it hard to
debug problems in this code path, as it did for me recently. Update the
test to use a standard assertion as well as asserting that the assertion
must have been called.

Signed-off-by: John Mulligan <jmulligan@redhat.com>