rgw: fix cloud-sync multi-tenancy scenario
At the moment, we cannot set buckets prefixed with tenant ID in the
`source_bucket` field from cloud-sync profiles (non-trivial config):
https://docs.ceph.com/en/latest/radosgw/cloud-sync-module/#non-trivial-configuration

This is because the `do_find_profile` function only searches in the
profiles configured using `bucket.name`, and it ignores `bucket.tenant`.

This is problematic in the RGW multi-tenancy scenario:
https://docs.ceph.com/en/latest/radosgw/multitenancy/#rgw-multi-tenancy

At the moment, we can only configure bucket name in the profile
`source_bucket` field. In the multi-tenancy scenario, this would sync
all the buckets (from all the tenants).

Without this fix, we cannot configure a cloud-sync profile that syncs
all the buckets from a tenant to a particular S3 target.

For example, we cannot do this:

* `tenantA/test-bucket` -> S3 target A
* `tenantB/test-bucket` -> S3 target B
* `tenantC/test-bucket` -> S3 target C

We can only do this at the moment:

* `test-bucket` -> S3 target A

If `test-bucket` is present in both `tenantA` and `tenantB`, both
buckets will be synced to S3 target A.

The idea would be to be able to do this:

* `tenantA/*` -> S3 target A
* `tenantB/*` -> S3 target B
* `tenantC/*` -> S3 target C

If `test-bucket` is present in all tenants, each tenant bucket is
synced to its own S3 target.

Fixes: https://tracker.ceph.com/issues/63395

Signed-off-by: Ionut Balutoiu <ibalutoiu@cloudbasesolutions.com>