osd/scrub: maintain a set of remote reservations
... instead of a simple counter.

This - as a preparation for the next commit, which will decouple
the "being reserved" state from the handling of scrub requests.
The planned changes to the scrub state machine will make
it harder to know when to clear the "being reserved" state.
The changes here would allow us to err on the side of caution,
i.e. trying to "un-count" a remote reservation even if it was not
actually reserved or was already deleted.

Signed-off-by: Ronen Friedman <rfriedma@redhat.com>