mds: allow lock state to be LOCK_MIX_SYNC in replica for filelock
When the auth MDS is transmitting state from LOCK_MIX to LOCK_SYNC
it will transmit the state to a intermidiate state LOCK_MIX_SYNC.
Then notifies the replicate MDSs and waits for the acks. But just
in case for some reasons the replicate MDSs couldn't successfully
transmit the state from LOCK_MIX to LOCK_MIX_SYNC and then
LOCK_MIX_SYNC2. And just the auth MDS failover or restart and
then resend the notification.

Fixes: https://tracker.ceph.com/issues/54833
Signed-off-by: Xiubo Li <xiubli@redhat.com>