src/mon/OSDMonitor.cc: more descriptive loggings for crc mismatch
Problem:
When mons are running different versions, e.g.,
during upgrades, we sometimes will encounter a crc
mismatch between the crc that is generated
from the leader MON (version n+1) and
the peon MON (version). Which will
later then gets resolved when all the MONs
are at version n+1.

Solution:
Provide a more informative loggings when
we encounter a CRC mismatch and also
output the mon versions so it's easier
for us to detect which if the crc mismatch
comes from OSDMap running mixed versions or
not.

Fixes: https://tracker.ceph.com/issues/63389

Signed-off-by: Kamoltat <ksirivad@redhat.com>